consolidate ts into single file, add parcel to serve locally
