+++ package.json
@@ -8,7 +8,7 @@
     "dev": "npx supervisor server.ts",
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "tsc",
-    "start": "ts-node-dev --respawn ./server.js"
+    "start": "parcel ./index.html"
   },
   "repository": {
     "type": "git",
@@ -17,7 +17,7 @@
   "author": "Samuel Barrett",
   "license": "ISC",
   "dependencies": {
-    "lodash": "^4.17.21",
+    "parcel": "^2.8.0",
     "spotify-web-api-node": "^5.0.2",
     "superagent": "^6.1.0"
   },
@@ -30,7 +30,6 @@
     "ts-node": "^10.2.1",
     "ts-node-dev": "^1.1.8",
     "typescript": "^4.5.2",
-    "webpack": "^5.75.0",
     "webpack-cli": "^5.0.0"
   }
 }
